pyproject.toml
[tool.poetry]
name = "lectureproject-django"
version = "0.1.0"
description = "강의사이트 제작"
authors = ["백관열 <djm030@naver.com>"]
readme = "README.md"
packages = [{include = "lectureproject_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pylint = "^2.16.2"
strawberry-graphql = {extras = ["debug-server"], version = "^0.159.0"}
django-cors-headers = "^3.14.0"
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
